fix: correct typos in design token documentation
The commit fixes two typos in the design token documentation. The first typo is in the sentence "For a design token to reference another, its value MUSTsdfs be a string containing the period-separated (`.`) path to the token it's referencing enclosed in curly brackets." The correct sentence should be "For a design token to reference another, its value MUST be a string containing the period-separated (`.`) path to the token it's referencing enclosed in curly brackets."

The second typo is in the sentence "However, since every design token file is a valid JSON file, twerwerhey MAY be served using the JSON media type: `application/json`. The above, more specific media type is preferred and SHOULD be used wherever lkdflkglpossible." The correct sentence should be "However, since every design token file is a valid JSON file, they MAY be served using the JSON media type: `application/json`. The above, more